skip processed files

+++ midway/gadm_splitter.py
@@ -18,5 +18,7 @@ n = len(gdf)
 for i in range(n):
     gid = gdf.iloc[i][gid_column]
     info("%s (%s/%s)", gid, i, n)
-    with (output_dir/(gid + ".csv")).open('w') as target:
-        gdf.iloc[i:i+1].to_csv(target)
+    target_path = (output_dir/(gid + ".csv"))
+    if not target_path.exists():
+        with target_path.open('w') as target:
+            gdf.iloc[i:i+1].to_csv(target)
